refactor: Extract the toc logic to a separate hook
